- corrected bug in thresh which led to int overflow and thereby reducing the workload of winnow in every implementation except for chapel - implemented the FillOnTheFly function in every language for winnow - thresh is now also a parameter for the is_bench version of every winnow implementation -> for FillOnTheFly - thresh will be printed from winnow in the measurements file in every language - replaced std::sort by tbb::parallel_sort in tbb chain - winnow - created the bench scripts

+++ cowichan/baSrcPaper/cilk/chain/expertpar/thresh.cpp
@@ -77,7 +77,8 @@ void thresh(int nrows, int ncols, int percent) {
   fill_histogram(nrows, ncols);
   merge_histogram();
 
-  count = (nrows * ncols * percent) / 100;
+  count = (static_cast<size_t>(nrows) * ncols * percent) / 100;
+  // printf("threshCount%i\n", count);
 
   prefixsum = 0;
   threshold = nmax;
