+++ cowichan/baSrcPaper/tbb/chain/expertpar/thresh.cc
@@ -81,7 +81,8 @@ void thresh(int nrows, int ncols, int percent) {
         }
       });
 
-  int count = (nrows * ncols * percent) / 100;
+  int count = (static_cast<size_t>(nrows) * ncols * percent) / 100;
+  // printf("thresh:%i\n",count);
 
   int prefixsum = 0;
   int threshold = nmax;
